fix: Move statsd parsing to metrics module
